+++ CHANGELOG.md
@@ -22,6 +22,7 @@ behavior is more as one would expect and hopefully the change won't cause any tr
 - SmartResize will automatically resize if the itemsProperty of the TableView changes value
 - Workspace.showHeadingLabelProperty controls whether the heading is shown in the Workspace toolbar or not
 - TableView/TreeTableView requestResize() will reapply SmartResize policy, useful after content change
+- Column.makeEditable() works for all number subclasses
 
 ### Fixed